Rewrite project to header files
I wanted to give C++ 20 modules a try, the idea is really good and I thought 3 years would be enough for them to work fine. It's not enough. I tried working on this project in both CLion and Visual Studio and in both of them, intelli sense and other IDE features didn't work half the time.